snn/winner UInt8 in snn node

+++ python_snn_node/python_snn_node/snn_node.py
@@ -268,7 +268,7 @@ class SNNNode(Node):
         self.create_subscription(Range, '/ultrasonic/front/scan', self._on_scan, qos_sensor) # Endre Range til LaserScan ved Gazebo
         self.create_subscription(Int32, '/snn/correct_output', self.cb_correct, 10)
 
-        self.pub_winner = self.create_publisher(Int32, '/snn/winner', 10)
+        self.pub_winner = self.create_publisher(UInt8, '/snn/winner', 10)
         self.pub_spikes = self.create_publisher(Int32MultiArray, '/snn/spikes', 10)
         self.pub_decision = self.create_publisher(String, '/snn/decision', 10)
